Fix Strava login state cookie handling in route handler
